+++ stories/components/UseSwipe.story.tsx
@@ -1,29 +1,20 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable react-hooks/rules-of-hooks */
-import { Button, Text } from '@chakra-ui/react'
+import { Box, Button, Text } from '@chakra-ui/react'
 import type { Meta, Story } from '@storybook/preact'
 import clsx from 'clsx'
 import { useMemo } from 'preact/hooks'
 
-import type { SwipeProps } from '@/components/swipe'
-import Swipe from '@/components/swipe'
+import UseSwipe from '@/components/UseSwipe'
 import { useBoolean } from '@/useBoolean'
 const threshold = -100
 
-const Template: Story<SwipeProps> = (args) => {
+const Template: Story = (args) => {
   const [isExceed, { on, off }] = useBoolean()
   return (
     <>
-      {isExceed && (
-        <Button
-          onClick={() => {
-            off()
-          }}
-        >
-          Reset
-        </Button>
-      )}
-      <Swipe
+      {isExceed && <Button onClick={off}>Reset</Button>}
+      <UseSwipe<HTMLDivElement>
         onSwipeEnd={(_, { reset, lengthY }) => {
           if (lengthY < threshold) {
             on()
@@ -32,7 +23,7 @@ const Template: Story<SwipeProps> = (args) => {
         }}
         {...args}
       >
-        {({ isSwiping, lengthY }) => {
+        {({ isSwiping, lengthY, ref }) => {
           const translateY = useMemo(
             () =>
               isExceed
@@ -43,7 +34,8 @@ const Template: Story<SwipeProps> = (args) => {
 
           const isEx = useMemo(() => lengthY < threshold, [lengthY])
           return (
-            <div
+            <Box
+              overflow="hidden"
               className={clsx(
                 'relative border flex flex-col rounded-2xl shadow h-screen',
                 {
@@ -55,16 +47,21 @@ const Template: Story<SwipeProps> = (args) => {
                 transform: translateY
               }}
             >
-              <div className="text-center">
-                <span className="w-24 h-1 inline-block rounded-full bg-gray-200" />
-              </div>
-              <div className="flex-1 grid place-items-center">
+              <Box
+                bg="gray.50"
+                borderBottom="1px"
+                borderColor="gray.200"
+                ref={ref}
+                className="text-center"
+              >
+                <span className="w-24 h-1 inline-block animate-pulse rounded-full bg-gray-200" />
                 <Text>Swipe down</Text>
-              </div>
-            </div>
+              </Box>
+              <div className="flex-1 grid place-items-center"></div>
+            </Box>
           )
         }}
-      </Swipe>
+      </UseSwipe>
     </>
   )
 }
@@ -72,8 +69,8 @@ const Template: Story<SwipeProps> = (args) => {
 export const Demo = Template.bind({})
 
 export default {
-  title: 'component/Swipe',
-  component: Swipe
-} as Meta<typeof Swipe>
+  title: 'component/UseSwipe',
+  component: UseSwipe
+} as Meta<typeof UseSwipe>
 
-export { Swipe }
+export { UseSwipe }