+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [v1.0.0-beta.40](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.39...v1.0.0-beta.40) (2021-10-19)
+
+## 🐛 Bug Fixes
+- [`3f435b4`](https://github.com/TomokiMiyauci/react-hookable/commit/3f435b4)   Change output declaration files by tsc
+
+# [v1.0.0-beta.39](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.38...v1.0.0-beta.39) (2021-10-19)
+
+## ✨ New Features
+- [`72dfa49`](https://github.com/TomokiMiyauci/react-hookable/commit/72dfa49)   Add &#x60;UseSwipe&#x60; for reactive swipe state wrapper
+
 # [v1.0.0-beta.38](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.37...v1.0.0-beta.38) (2021-10-19)
 
 ## ✨ New Features